+++ backends/xla/gen_standard_ops.go
@@ -334,31 +334,36 @@ func (b *Builder) Conj(x backends.Op) (backends.Op, error) {
 	return xla_result, nil
 }
 
-// ConvGeneralDilated is a generic Convolution operation offered by XLA.
-// featureAxisAfter defines whether the features (aka. channels or depth) axis comes after the
-// spatial dimension. Example: a 2D input can be one of the two:
-//   - featureAxisAfter=false: input=[batch_size, features, height, width], filter=[output_features, input_features, height, width]
-//   - featureAxisAfter=true:  input=[batch_size, height, width, features], filter=[output_features, height, width, input_features]
-//
+// ConvGeneral is a generic Convolution operation with support for:
+// - Arbitrary number of spatial axes.
+// - Arbitrary transposition of axes.
+// - Strides and padding.
+// - Dilations of the input.
+// - Dilations of the kernel, aka. atrous convolution.
+// - Channels grouping (on the input channels).
+// - Batch grouping.
 // Some details in https://www.tensorflow.org/xla/operation_semantics#convwithgeneralpadding_convolution.
 // There operand and filter are called lhs and rhs.
 // (XLA documentation is unfortunately poor, much is guess-work).
 // Also useful, https://arxiv.org/pdf/1603.07285v1.pdf.
-func (b *Builder) ConvGeneralDilated(operand, filter backends.Op, axes backends.ConvolveAxesConfig, strides []int, paddings [][2]int, inputDilation, filterDilation []int, filterGroupCount, batchGroupCount int) (backends.Op, error) {
-	xla_operand, err := b.verifyAndCastOp(operand, "operand")
+// Note:
+//   - Another common term for "channels" is "features".
+//   - "Kernel" is also commonly called "weights" or "filters".
+func (b *Builder) ConvGeneral(input, kernel backends.Op, axes backends.ConvolveAxesConfig, strides []int, paddings [][2]int, inputDilations, kernelDilations []int, channelGroupCount, batchGroupCount int) (backends.Op, error) {
+	xla_input, err := b.verifyAndCastOp(input, "input")
 	if err != nil {
-		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneralDilated", BackendName)
+		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneral", BackendName)
 	}
 
-	xla_filter, err := b.verifyAndCastOp(filter, "filter")
+	xla_kernel, err := b.verifyAndCastOp(kernel, "kernel")
 	if err != nil {
-		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneralDilated", BackendName)
+		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneral", BackendName)
 	}
 
 	xla_axes := convertConvolveAxesConfig(axes)
-	xla_result, err := xlabuilder.ConvGeneralDilated(xla_operand, xla_filter, xla_axes, strides, paddings, inputDilation, filterDilation, filterGroupCount, batchGroupCount)
+	xla_result, err := xlabuilder.ConvGeneral(xla_input, xla_kernel, xla_axes, strides, paddings, inputDilations, kernelDilations, channelGroupCount, batchGroupCount)
 	if err != nil {
-		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneralDilated", BackendName)
+		return nil, errors.WithMessagef(err, "Backend %q: failed ConvGeneral", BackendName)
 	}
 	return xla_result, nil
 }